extract method that fetches product catalogue
